Move NativeCodeLoader to top level package so that it can be made package private
